fix bug

+++ app/models.py
@@ -1,5 +1,5 @@
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String, JSON
+from sqlalchemy import Column, Integer, BigInteger, String, JSON
 from sqlalchemy import create_engine
 from envparse import Env
 
@@ -17,8 +17,8 @@ class Company(Base):
     id = Column(Integer, autoincrement=True)
     name = Column(String(100), primary_key=True)
     code = Column(String, nullable=True)
-    inn = Column(Integer, nullable=True)
-    kpp = Column(Integer, nullable=True)
+    inn = Column(BigInteger, nullable=True)
+    kpp = Column(BigInteger, nullable=True)
     address = Column(JSON, nullable=True)
